Fix two compiler warnings.
Thanks to Mark Harris for the report.

+++ src/info.c
@@ -115,7 +115,7 @@ static int op_tags_ensure_capacity(OpusTags *_tags,size_t _ncomments){
   /*We only support growing.
     Trimming requires cleaning up the allocated strings in the old space, and
      is best handled separately if it's ever needed.*/
-  OP_ASSERT(_ncomments>=cur_ncomments);
+  OP_ASSERT(_ncomments>=(size_t)cur_ncomments);
   comment_lengths=_tags->comment_lengths;
   comment_lengths=(int *)_ogg_realloc(_tags->comment_lengths,size);
   if(OP_UNLIKELY(comment_lengths==NULL))return OP_EFAULT;
